NPM Implementation (#338)
Honestly. I hate NPM

I am going to rant in the commit message
When NPM requests the file. This is what it gives you
`@shoppence/shoppence-ts/-/@shoppence/shoppence-ts-0.1.1.tgz`
So `{project_name}/-/{project_name}-{version}.tgz` Now in the dist tags in the package.json you can specify the tarball. But NPM does not care what you set it to. It always requests this.

Now this file location format is just horrible for organization. 

So I override this data a bit. It is barely an improvement `@shoppence/shoppence-ts/0.1.1/@shoppence/shoppence-ts-0.1.1.tgz` So I can keep all of the version files in one location.

I do understand that I lack knowledge on how NPM works. However, it is still a pain and I truly hate it.

Also NPM publish uses BASE64 inside JSON for sending the file. WTF